change path and refactor the code
