+++ magentacloud25/custom/apps/files_sharing/src/components/SharingEntryLink.vue
@@ -289,11 +289,11 @@ import NcActionSeparator from '@nextcloud/vue/dist/Components/NcActionSeparator'
 import NcActionTextEditable from '@nextcloud/vue/dist/Components/NcActionTextEditable'
 import NcActions from '@nextcloud/vue/dist/Components/NcActions'
 import NcAvatar from '@nextcloud/vue/dist/Components/NcAvatar'
-import ExternalShareAction from '../../../../../../../release25.0.6/apps/files_sharing/src/components/ExternalShareAction.vue'
-import SharePermissionsEditor from '../../../../../../../release25.0.6/apps/files_sharing/src/SharePermissionsEditor.vue'
-import GeneratePassword from '../../../../../../../release25.0.6/apps/files_sharing/src/utils/GeneratePassword.js'
-import Share from '../../../../../../../release25.0.6/apps/files_sharing/src/models/Share.js'
-import SharesMixin from '../../../../../../../release25.0.6/apps/files_sharing/src/mixins/SharesMixin.js'
+import ExternalShareAction from '../../../../../../../nextcloud/apps/files_sharing/src/components/ExternalShareAction.vue'
+import SharePermissionsEditor from '../../../../../../../nextcloud/apps/files_sharing/src/SharePermissionsEditor.vue'
+import GeneratePassword from '../../../../../../../nextcloud/apps/files_sharing/src/utils/GeneratePassword.js'
+import Share from '../../../../../../../nextcloud/apps/files_sharing/src/models/Share.js'
+import SharesMixin from '../../../../../../../nextcloud/apps/files_sharing/src/mixins/SharesMixin.js'
 export default {
 	name: 'SharingEntryLink',
 	components: {
