+++ magentacloud25/custom/apps/files_sharing/src/views/SharingTab.vue
@@ -90,16 +90,16 @@ import { generateOcsUrl } from '@nextcloud/router'
 import NcAvatar from '@nextcloud/vue/dist/Components/NcAvatar'
 import axios from '@nextcloud/axios'
 import { loadState } from '@nextcloud/initial-state'
-import Config from '../../../../../../../release25.0.6/apps/files_sharing/src/services/ConfigService'
-import { shareWithTitle } from '../../../../../../../release25.0.6/apps/files_sharing/src/utils/SharedWithMe'
-import Share from '../../../../../../../release25.0.6/apps/files_sharing/src/models/Share'
-import ShareTypes from '../../../../../../../release25.0.6/apps/files_sharing/src/mixins/ShareTypes'
-import SharingEntryInternal from '../../../../../../../release25.0.6/apps/files_sharing/src/components/SharingEntryInternal'
-import SharingEntrySimple from '../../../../../../../release25.0.6/apps/files_sharing/src/components/SharingEntrySimple'
+import Config from '../../../../../../../nextcloud/apps/files_sharing/src/services/ConfigService'
+import { shareWithTitle } from '../../../../../../../nextcloud/apps/files_sharing/src/utils/SharedWithMe'
+import Share from '../../../../../../../nextcloud/apps/files_sharing/src/models/Share'
+import ShareTypes from '../../../../../../../nextcloud/apps/files_sharing/src/mixins/ShareTypes'
+import SharingEntryInternal from '../../../../../../../nextcloud/apps/files_sharing/src/components/SharingEntryInternal'
+import SharingEntrySimple from '../../../../../../../nextcloud/apps/files_sharing/src/components/SharingEntrySimple'
 import SharingInput from '../components/SharingInput'
-import SharingInherited from '../../../../../../../release25.0.6/apps/files_sharing/src/views/SharingInherited'
-import SharingLinkList from '../../../../../../../release25.0.6/apps/files_sharing/src/views/SharingLinkList'
-import SharingList from '../../../../../../../release25.0.6/apps/files_sharing/src/views/SharingList.vue'
+import SharingInherited from '../../../../../../../nextcloud/apps/files_sharing/src/views/SharingInherited'
+import SharingLinkList from '../../../../../../../nextcloud/apps/files_sharing/src/views/SharingLinkList'
+import SharingList from '../../../../../../../nextcloud/apps/files_sharing/src/views/SharingList.vue'
 export default {
 	name: 'SharingTab',
 	components: {
